refactor: align types with merged PRs #23 and #24
- Remove duplicate VideoInfo and AdjacentEpisodeVideos types from
  fetch-data.ts; import from @/types instead
- Update videos-player and simple-videos-player to import VideoInfo
  from @/types instead of fetch-data
- Replace manual Number() BigInt conversions with bigIntToNumber()
  from @/utils/typeGuards for consistency
- Replace magic-string SERIES_DELIM in urdf-viewer with
  CHART_CONFIG.SERIES_NAME_DELIMITER from @/utils/constants
