+++ src/components/videos-player.tsx
@@ -3,7 +3,7 @@
 import { useEffect, useRef, useState } from "react";
 import { useTime } from "../context/time-context";
 import { FaExpand, FaCompress, FaTimes, FaEye } from "react-icons/fa";
-import type { VideoInfo } from "@/app/[org]/[dataset]/[episode]/fetch-data";
+import type { VideoInfo } from "@/types";
 
 type VideoPlayerProps = {
   videosInfo: VideoInfo[];
